Started sketching LUA embedding.
There is a desperate need to have resource-specific loader
implementation in the resource tree. Simply, for example, having a piece
of code creating skybox cubemap texture from provided bitmaps. This
could be done by inserting D source pieces to directory tree and
compiling them inside the executable, but for common maintenance I have
thought to have loader scripts which are stored in the BLOB with the
resources. For this purpose, I have been thinking of embedding LUA in
the engine - and, of course, making this enables a whole new world to
experiment!

For now, LUA is added as statically linked library, and there is only
Linux 64-bit version in the directory tree at the moment.