ERRORIF() updated.
Previous version of ERRORIF took in a value, checked it against zero,
and returned it. This way it was possible to insert it in to equations.
The behaviour was just too confusing, and it was now changed to take in
a boolean value indicating error (true => error). This still works the
opposite way compared to assert(), but it is acceptable: we check for
errors, not validate values.